# BUG: .transform(...) with "first" and "last" fail when axis=1